[mobile][photos] Separate ML DB facade from Dart implementation
